+++ pkg/interop/server_authn.go
@@ -16,41 +16,19 @@ package interop
 
 import (
 	"context"
-	"crypto/tls"
 	"encoding/json"
 	"net"
 	"net/http"
 	"net/url"
 	"strings"
 
+	"go.thethings.network/lorawan-stack/v3/pkg/log"
 	"go.thethings.network/lorawan-stack/v3/pkg/types"
+	"gopkg.in/square/go-jose.v2/jwt"
 )
 
-func (s *Server) verifySenderCertificate(ctx context.Context, senderID string, state *tls.ConnectionState) (addrs []string, err error) {
-	// TODO: Support reading TLS client certificate from proxy headers (https://github.com/TheThingsNetwork/lorawan-stack/issues/717).
-	senderClientCAs, err := s.SenderClientCAs(ctx, senderID)
-	if err != nil {
-		return nil, err
-	}
-	for _, chain := range state.VerifiedChains {
-		peerCert, clientCA := chain[0], chain[len(chain)-1]
-		for _, senderClientCA := range senderClientCAs {
-			if clientCA.Equal(senderClientCA) {
-				// If the TLS client certificate contains DNS addresses, use those. Otherwise, fallback to using CommonName as address.
-				if len(peerCert.DNSNames) > 0 {
-					addrs = append([]string(nil), peerCert.DNSNames...)
-				} else {
-					addrs = []string{peerCert.Subject.CommonName}
-				}
-				return
-			}
-		}
-	}
-	// TODO: Verify state.PeerCertificates[0] with senderClientCAs as Roots and state.PeerCertificates[1:] as Intermediates.
-	// (https://github.com/TheThingsNetwork/lorawan-stack/issues/718).
-	return nil, errUnauthenticated.New()
-}
-
+// verifySenderNSID verifies that one of the address patterns matches the NSID.
+// The pattern may contain a wildcard (*.host) or port (host:1885).
 func verifySenderNSID(patterns []string, nsID string) error {
 	if len(patterns) == 0 {
 		return errCallerNotAuthorized.WithAttributes("target", nsID)
@@ -88,16 +66,29 @@ nextPattern:
 }
 
 type authInfo interface {
-	GetAddresses() []string
+	addressPatterns() []string
 }
 
-// NetworkServerAuthInfo contains authentication information of the Network Server.
+// NetworkServerAuthInfo contains the authentication information of a Network Server.
 type NetworkServerAuthInfo struct {
 	NetID     types.NetID
 	Addresses []string
 }
 
-func (n NetworkServerAuthInfo) GetAddresses() []string { return n.Addresses }
+func (n NetworkServerAuthInfo) addressPatterns() []string { return n.Addresses }
+
+// Require returns an error if the given NetID does not match, or if the NSID is not matched by an address pattern.
+func (n NetworkServerAuthInfo) Require(netID types.NetID, nsID *string) error {
+	if !n.NetID.Equal(netID) {
+		return errUnauthenticated.New()
+	}
+	if nsID != nil {
+		if err := verifySenderNSID(n.Addresses, *nsID); err != nil {
+			return err
+		}
+	}
+	return nil
+}
 
 type nsAuthInfoKeyType struct{}
 
@@ -105,53 +96,23 @@ var nsAuthInfoKey nsAuthInfoKeyType
 
 // NewContextWithNetworkServerAuthInfo returns a derived context with the given authentication information of the
 // Network Server.
-func NewContextWithNetworkServerAuthInfo(parent context.Context, authInfo NetworkServerAuthInfo) context.Context {
+func NewContextWithNetworkServerAuthInfo(parent context.Context, authInfo *NetworkServerAuthInfo) context.Context {
 	return context.WithValue(parent, nsAuthInfoKey, authInfo)
 }
 
 // NetworkServerAuthInfoFromContext returns the authentication information of the Network Server from context.
-func NetworkServerAuthInfoFromContext(ctx context.Context) (NetworkServerAuthInfo, bool) {
-	authInfo, ok := ctx.Value(nsAuthInfoKey).(NetworkServerAuthInfo)
+func NetworkServerAuthInfoFromContext(ctx context.Context) (*NetworkServerAuthInfo, bool) {
+	authInfo, ok := ctx.Value(nsAuthInfoKey).(*NetworkServerAuthInfo)
 	return authInfo, ok
 }
 
-func (s *Server) authenticateNS(ctx context.Context, r *http.Request, data []byte) (context.Context, error) {
-	var header NsMessageHeader
-	if err := json.Unmarshal(data, &header); err != nil {
-		return nil, err
-	}
-	if !header.ProtocolVersion.SupportsNSID() && header.SenderNSID != nil {
-		return nil, ErrMalformedMessage.New()
-	}
-
-	// If the client presents a TLS client certificate, use that for authentication.
-	if state := r.TLS; state != nil && len(state.PeerCertificates) > 0 {
-		addrs, err := s.verifySenderCertificate(ctx, types.NetID(header.SenderID).String(), state)
-		if err != nil {
-			return nil, err
-		}
-		if header.SenderNSID != nil {
-			if err := verifySenderNSID(addrs, *header.SenderNSID); err != nil {
-				return nil, err
-			}
-		}
-		return NewContextWithNetworkServerAuthInfo(ctx, NetworkServerAuthInfo{
-			NetID:     types.NetID(header.SenderID),
-			Addresses: addrs,
-		}), nil
-	}
-
-	// TODO: Verify Packet Broker token (https://github.com/TheThingsNetwork/lorawan-stack/issues/4703).
-
-	return ctx, nil
-}
-
+// ApplicationServerAuthInfo contains the authentication information of an Application Server.
 type ApplicationServerAuthInfo struct {
 	ASID      string
 	Addresses []string
 }
 
-func (a ApplicationServerAuthInfo) GetAddresses() []string { return a.Addresses }
+func (a ApplicationServerAuthInfo) addressPatterns() []string { return a.Addresses }
 
 type asAuthInfoKeyType struct{}
 
@@ -159,31 +120,113 @@ var asAuthInfoKey asAuthInfoKeyType
 
 // NewContextWithApplicationServerAuthInfo returns a derived context with the given authentication information of the
 // Application Server.
-func NewContextWithApplicationServerAuthInfo(parent context.Context, authInfo ApplicationServerAuthInfo) context.Context {
+func NewContextWithApplicationServerAuthInfo(parent context.Context, authInfo *ApplicationServerAuthInfo) context.Context {
 	return context.WithValue(parent, asAuthInfoKey, authInfo)
 }
 
 // ApplicationServerAuthInfoFromContext returns the authentication information of the Application Server from context.
-func ApplicationServerAuthInfoFromContext(ctx context.Context) (ApplicationServerAuthInfo, bool) {
-	authInfo, ok := ctx.Value(asAuthInfoKey).(ApplicationServerAuthInfo)
+func ApplicationServerAuthInfoFromContext(ctx context.Context) (*ApplicationServerAuthInfo, bool) {
+	authInfo, ok := ctx.Value(asAuthInfoKey).(*ApplicationServerAuthInfo)
 	return authInfo, ok
+}
+
+type tokenVerifier interface {
+	VerifyNetworkServer(context.Context, *jwt.JSONWebToken) (*NetworkServerAuthInfo, error)
+	VerifyApplicationServer(context.Context, *jwt.JSONWebToken) (*ApplicationServerAuthInfo, error)
+}
+
+// authenticateNS authenticates the client as a Network Server.
+//
+// If the client presents a TLS client certificate, it is verified against the trusted CAs of the NetID.
+// Any DNS names in the X.509 Subject Alternative Names are taken as address patterns used to verify the NSID. If there
+// are no DNS names, the Common Name is used as the single address pattern.
+// If the TLS client certificate verification fails, this method returns an error.
+//
+// If the client presents a Bearer token in the Authorization header of the HTTP request, it is verified as token issued
+// by Packet Broker.
+func (s *Server) authenticateNS(ctx context.Context, r *http.Request, data []byte) (context.Context, error) {
+	var header NsMessageHeader
+	if err := json.Unmarshal(data, &header); err != nil {
+		return nil, ErrMalformedMessage.WithCause(err)
+	}
+	if !header.ProtocolVersion.SupportsNSID() && header.SenderNSID != nil {
+		return nil, ErrMalformedMessage.New()
+	}
+
+	for _, authFunc := range []func(context.Context) (*NetworkServerAuthInfo, error){
+		// Verify TLS client certificate.
+		func(ctx context.Context) (*NetworkServerAuthInfo, error) {
+			if r.TLS == nil || len(r.TLS.PeerCertificates) == 0 {
+				return nil, nil
+			}
+			addrs, err := s.verifySenderCertificate(ctx, types.NetID(header.SenderID).String(), r.TLS)
+			if err != nil {
+				return nil, err
+			}
+			return &NetworkServerAuthInfo{
+				NetID:     types.NetID(header.SenderID),
+				Addresses: addrs,
+			}, nil
+		},
+		// Verify token in a best-effort manner.
+		func(ctx context.Context) (*NetworkServerAuthInfo, error) {
+			logger := log.FromContext(ctx).WithField("authenticator", "packetbroker")
+			authz := strings.SplitN(r.Header.Get("Authorization"), " ", 2)
+			if len(authz) < 2 || strings.ToLower(authz[0]) != "bearer" {
+				return nil, nil
+			}
+			token, err := jwt.ParseSigned(authz[1])
+			if err != nil {
+				logger.WithError(err).Debug("Failed to parse token")
+				return nil, nil
+			}
+			var claims jwt.Claims
+			err = token.UnsafeClaimsWithoutVerification(&claims)
+			if err != nil {
+				logger.WithError(err).Debug("Failed to parse claims")
+				return nil, nil
+			}
+			if tokenVerifier, ok := s.tokenVerifiers[claims.Issuer]; ok {
+				authInfo, err := tokenVerifier.VerifyNetworkServer(ctx, token)
+				if err != nil {
+					return nil, errUnauthenticated.WithCause(err)
+				}
+				return authInfo, nil
+			}
+			logger.WithError(err).WithField("issuer", claims.Issuer).Debug("Unknown token issuer")
+			return nil, nil
+		},
+	} {
+		authInfo, err := authFunc(ctx)
+		if err != nil {
+			return nil, ErrUnknownSender.WithCause(err)
+		}
+		if authInfo != nil {
+			if err := authInfo.Require(types.NetID(header.SenderID), header.SenderNSID); err != nil {
+				return nil, ErrUnknownSender.WithCause(err)
+			}
+			return NewContextWithNetworkServerAuthInfo(ctx, authInfo), nil
+		}
+	}
+
+	return nil, ErrUnknownSender.New()
 }
 
 func (s *Server) authenticateAS(ctx context.Context, r *http.Request, data []byte) (context.Context, error) {
 	var header AsMessageHeader
 	if err := json.Unmarshal(data, &header); err != nil {
-		return nil, err
+		return nil, ErrMalformedMessage.WithCause(err)
 	}
 
 	state := r.TLS
 	if state == nil {
-		return nil, errUnauthenticated.New()
+		return nil, ErrUnknownSender.New()
 	}
 	addrs, err := s.verifySenderCertificate(ctx, header.SenderID, state)
 	if err != nil {
-		return nil, err
+		return nil, ErrUnknownSender.WithCause(err)
 	}
-	return NewContextWithApplicationServerAuthInfo(ctx, ApplicationServerAuthInfo{
+	return NewContextWithApplicationServerAuthInfo(ctx, &ApplicationServerAuthInfo{
 		ASID:      header.SenderID,
 		Addresses: addrs,
 	}), nil